fix: fix render problems caused by rtkq cash
